[Misc] Simply PoolerOutput and move to `v1/outputs`
